processor/requirements.txt
python-dotenv
requests
llama-cpp-python
llama-cpp-agent
openai
pydantic
google-genai
